[Refactor] Refactor pipeline RF core types out of observer-backed runtime types
